Face api data log to console on mobile
Facial expressions are being tracked and I can at the same time also track touch events and mobile sensor events. The table export has an issue with the date and time column. The stream video of the webgazer is there but it is not tracking eye gaze at the moment.